Fix Phantom skills, fix Util.getRandom(start, end) not including end

+++ src/main/java/net/swordie/ms/connection/packet/UserLocal.java
@@ -138,7 +138,7 @@ public class UserLocal {
         return outPacket;
     }
 
-    public static OutPacket incJudgementStack(byte amount) {
+    public static OutPacket incJudgementStack(int amount) {
         OutPacket outPacket = new OutPacket(OutHeader.INC_JUDGEMENT_STACK_RESPONSE);
 
         outPacket.encodeByte(amount);
